fix: fixed metrics, moved metrics component to shared module
