+++ src/app/shared/components/metrics/metrics.component.html
