+++ src/app/shared/components/metrics/metrics.component.scss
