+++ src/app/youtube/pages/detailed-information-page/detailed-information-page.component.html
@@ -1,4 +1,4 @@
-<div class="detailed">
+<div class="detailed" *ngIf="currentItem">
     <app-custom-button
         additionalClasses="detailed__back-button"
         (click)="backToMain()"
@@ -8,30 +8,27 @@
 
     <div class="detailed__image">
         <img
-            [src]="currentItem?.snippet?.thumbnails?.standard?.url"
-            [alt]="currentItem?.snippet?.title"
+            [src]="currentItem.snippet.thumbnails.standard.url"
+            [alt]="currentItem.snippet.title"
         />
     </div>
 
     <div
         class="detailed__wrapper"
-        [appColorByDate]="currentItem?.snippet?.publishedAt || ''"
+        [appColorByDate]="currentItem.snippet.publishedAt || ''"
     >
         <div class="detailed__wrapper-titledate">
             <h1 class="detailed__wrapper-title">
-                {{ currentItem?.snippet?.title }}
+                {{ currentItem.snippet.title }}
             </h1>
             <div class="detailed__wrapper-date">
-                {{
-                    currentItem?.snippet?.publishedAt
-                        | date : "fullDate"
-                }}
+                {{ currentItem.snippet.publishedAt | date : "fullDate" }}
             </div>
         </div>
         <div class="detailed__wrapper-about">
             <div class="detailed__wrapper-subtitle">Description:</div>
             <div class="detailed__wrapper-description">
-                {{ currentItem?.snippet?.description }}
+                {{ currentItem.snippet.description }}
             </div>
         </div>
 
@@ -40,3 +37,5 @@
         </div>
     </div>
 </div>
+
+<div *ngIf="!currentItem" class="detailed__loading">Loading...</div>
